Unngå feil med når bruker skriver feilaktig fødselsdato og klikker på radioknapp
